chore: payment -> payment gateway 패키지 이름 변경
