--- src/main/java/com/hyoguoo/paymentplatform/payment/exception/common/PaymentExceptionHandler.java
@@ -1,30 +1,0 @@
-package com.hyoguoo.paymentplatform.payment.exception.common;
-
-import com.hyoguoo.paymentplatform.core.response.ErrorResponse;
-import com.hyoguoo.paymentplatform.payment.exception.PaymentFoundException;
-import lombok.extern.slf4j.Slf4j;
-import org.springframework.core.Ordered;
-import org.springframework.core.annotation.Order;
-import org.springframework.http.HttpStatus;
-import org.springframework.http.ResponseEntity;
-import org.springframework.web.bind.annotation.ExceptionHandler;
-import org.springframework.web.bind.annotation.RestControllerAdvice;
-
-@Order(Ordered.HIGHEST_PRECEDENCE)
-@Slf4j
-@RestControllerAdvice
-public class PaymentExceptionHandler {
-
-    @ExceptionHandler(PaymentFoundException.class)
-    public ResponseEntity<ErrorResponse> catchRuntimeException(PaymentFoundException e) {
-        log.warn(e.getMessage());
-
-        return ResponseEntity.status(HttpStatus.NOT_FOUND)
-                .body(ErrorResponse.of(
-                                e.getCode(),
-                                e.getMessage()
-                        )
-                );
-    }
-}
-
