+++ src/main/java/com/hyoguoo/paymentplatform/paymentgateway/presentation/dto/request/TossCancelClientRequest.java
@@ -1,0 +1,15 @@
+package com.hyoguoo.paymentplatform.paymentgateway.presentation.dto.request;
+
+import lombok.Builder;
+import lombok.Getter;
+import lombok.RequiredArgsConstructor;
+
+@Getter
+@Builder
+@RequiredArgsConstructor
+public class TossCancelClientRequest {
+
+    private final String paymentKey;
+    private final String cancelReason;
+    private final String idempotencyKey;
+}
